ai-pipeline/requirements.txt
flask
kafka-python
requests
joblib
scikit-learn